Tests and Fix for RCP-39. (http://opensource2.atlassian.com/projects/spring/browse/RCP-39)
Now all the form's commands are 'guarded' by the same object.
The same object can be used to attach custom 'Guarded' components.
Guarding is customisable per managed component and allows to be triggered by a mix of the distinct formmodel-states:
- being enabled itself
- having changes (isDirty)
- having no-validation-errors

Furthermore this will make the form's commands be guarded in the following way
- setting new form object enabled: as soon as the form is enabled
- reverting enabled: only if the form is enabled AND the form has changes to revert
- commiting enabled: only if the form is enables AND the form has changes to commit that are all valid.

Finally this also removes the former AbstractForm.setFormEnabledGuarded as the new implementation should be the prefered way of working and a deprecation path was not concidered as of now.


git-svn-id: https://spring-rich-c.svn.sourceforge.net/svnroot/spring-rich-c/trunk/spring-richclient@745 817809c7-9d0e-0410-b92d-a7ac8b8adc13